+++ OSLC4Net_SDK/OSLC4Net.Core.DotNetRdfProvider/DotNetRdfHelper.cs
@@ -20,7 +20,8 @@ using System.Numerics;
 using System.Reflection;
 using System.Xml.Linq;
 using CommunityToolkit.Diagnostics;
-using log4net;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Abstractions;
 using OSLC4Net.Core.Attribute;
 using OSLC4Net.Core.Exceptions;
 using OSLC4Net.Core.Model;
@@ -34,8 +35,14 @@ namespace OSLC4Net.Core.DotNetRdfProvider;
 /// <summary>
 ///     A class to assist with serialization and de-serialization of RDF/XML from/to .NET objects
 /// </summary>
-public static class DotNetRdfHelper
+public class DotNetRdfHelper(ILogger<DotNetRdfHelper> logger)
 {
+    public DotNetRdfHelper() : this(NullLoggerFactory.Instance.CreateLogger<DotNetRdfHelper>())
+    {
+    }
+
+    private readonly ILogger<DotNetRdfHelper> _logger =
+        logger;
     private const string PROPERTY_TOTAL_COUNT = "totalCount";
     private const string PROPERTY_NEXT_PAGE = "nextPage";
 
@@ -55,8 +62,6 @@ public static class DotNetRdfHelper
 
     private static readonly int METHOD_NAME_START_GET_LENGTH = METHOD_NAME_START_GET.Length;
     private static readonly int METHOD_NAME_START_IS_LENGTH = METHOD_NAME_START_IS.Length;
-
-    private static readonly ILog logger = LogManager.GetLogger(typeof(DotNetRdfHelper));
 
     /// <summary>
     ///     Create an RDF graph from a collection of .NET objects
@@ -235,7 +240,7 @@ public static class DotNetRdfHelper
     /// <param name="resource"></param>
     /// <param name="beanType"></param>
     /// <returns></returns>
-    public static object FromDotNetRdfNode(IUriNode resource,
+    public object FromDotNetRdfNode(IUriNode resource,
         IGraph? graph,
         Type beanType)
     {
@@ -262,7 +267,7 @@ public static class DotNetRdfHelper
     /// <param name="graph"></param>
     /// <param name="beanType"></param>
     /// <returns></returns>
-    public static object FromDotNetRdfGraph(IGraph graph,
+    public object FromDotNetRdfGraph(IGraph graph,
         Type beanType)
     {
         Type[] types = { beanType };
@@ -307,7 +312,7 @@ public static class DotNetRdfHelper
         return results;
     }
 
-    private static void FromDotNetRdfNode(
+    private void FromDotNetRdfNode(
         IDictionary<Type, IDictionary<string, MemberInfo>> typePropertyDefinitionsToSetMethods,
         Type beanType,
         object bean,
@@ -393,10 +398,10 @@ public static class DotNetRdfHelper
                 {
                     if (extendedProperties == null)
                     {
-                        logger.Warn("Set method not found for object type:  " +
-                                    beanType.Name +
-                                    ", uri:  " +
-                                    uri);
+                        _logger.LogWarning("Set method not found for object type:  " +
+                                           beanType.Name +
+                                           ", uri:  " +
+                                           uri);
                     }
                     else
                     {
@@ -1000,7 +1005,7 @@ public static class DotNetRdfHelper
         return candidatePrefix;
     }
 
-    private static object HandleExtendedPropertyValue(Type beanType,
+    private object HandleExtendedPropertyValue(Type beanType,
         INode obj,
         IGraph graph,
         IDictionary<string, object> visitedResources)